Rules: Add special attack abilities
This patch implements most of the special attack abilities.
This is done mostly via SpecialAbilityImpact, which, depending on
the chosen ability, modifies the roll mechanics of the attack,
parry, dodge and damage roll.
In the easiest cases, this is done by additive modifiers. For more
complex cases, though, the class provides slots for callbacks,
which by inversion of control from CombatRoll.roll() and damageRoll()
handle the roll mechanics by themselves.
Apart from thorough testing, this (mostly) closes issue #4.