UI: Preserve navigation stack upon playing transition animations
Dice animations now correctly preserve the navigation stack.
The fadeDice() API now takes a Widget? instead of a Widget for the
target screen, passing null here means returning to the original
screen.
Since the objective of #18 is to consistently display dice throw
results as popups anyway, this leaves room for some further
simplification.
This patch closes issues #27 and #28.