UI: Deactivate impossible action buttons
Actions with FWeff < 1 are by default now deactivated in the UI.
Note: Technically, there is no distinction between actions
disallowed by the rulebook (for instance dodging after using
special ability "Vorstoß", and things that are plain impossible
due to adverse states, the character being almost dead for example.
The former are implemented via a modifier of -20, so in theory
one would be able to "illegally" activate the button by setting
a sufficiently high GM modifier, but this seems unlikely enough
not to bother with the distinction.