Core: Make attribute rolls remember confirmation throws
To be fully auditable, we need to remember the results of confirmation
throws. To that end, AttributeRollResult now stores a DiceValue?
instead of an int? for the result.
DiceValue behaves like an integer in every respect, but carries an
additional int? confirmationThrow storing the result of the second
roll in case the first one is a 1 or a 20.