Bugfix: Display correct confirmation throw result
The result of the confirmation throw was not correctly displayed,
instead, we displayed the result of the original throw twice.
While this at the surface was just a simple typo, at heart lay a
confusion about the interface between DiceValue (which behaves
like an integer, optionally carrying the result of a confirmation
throw) and Dice (and its subclasses), which are responsible for
displaying the results of *single* dice throws, but used to carry
a DiceValue? as an attribute.
This patch cleans this up in the sense that now all subclasses of
Dice only carry an int? around, representing their actual result.